requirements.txt
requests==2.32.3
pandas==2.2.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pymysql==1.1.1
mysql-connector-python==8.4.0
SQLAlchemy==2.0.31
confluent-kafka==2.4.0
sentry-sdk[flask]==2.6.0
flask-restful==0.3.10
kafka-python==2.0.2
boto3==1.34.131
jaeger-client==4.8.0
Flask-OpenTracing==2.0.0
gunicorn==22.0.0
cryptography==42.0.8
numpy==1.26.4
psycopg2-binary==2.9.9
pylint==2.15.0
tensorflow
google-generativeai
matplotlib
datetime
plotly
PyYAML
scikit-learn
xgboost